requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
importlib-metadata==3.7.0
kombu==5.0.2
Pillow==8.1.0
pkg-resources==0.0.0
prompt-toolkit==3.0.16
pytz==2021.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
